+++ src/internal/packager/helm/destroy.go
@@ -7,7 +7,7 @@ package helm
 import (
 	"regexp"
 
-	"github.com/defenseunicorns/zarf/src/internal/cluster"
+	"github.com/defenseunicorns/zarf/src/pkg/cluster"
 	"github.com/defenseunicorns/zarf/src/pkg/message"
 	"helm.sh/helm/v3/pkg/action"
 )
